+++ azure/storage/file/_serialization.py
@@ -14,21 +14,10 @@
 #--------------------------------------------------------------------------
 from time import time
 from wsgiref.handlers import format_date_time
-from .._common_serialization import (
-    _parse_response_for_dict,
-    ETree,
-    _ETreeXmlToObject,
-)
-from .._common_conversion import (
-    _decode_base64_to_text,
-    _encode_base64,
-)
+from .._common_serialization import _parse_response_for_dict
 from .._serialization import _update_storage_header
 from .models import (
-    FileAndDirectoryEnumResults,
     FileResult,
-    File,
-    Directory
 )
 
 
@@ -41,28 +30,6 @@ def _update_storage_file_header(request, authentication):
     authentication.sign_request(request)
 
     return request.headers
-
-
-def _parse_file_enum_results_list(response):
-    respbody = response.body
-    return_obj = FileAndDirectoryEnumResults()
-    enum_results = ETree.fromstring(respbody)
-
-    for child in enum_results.findall('./Entries/File'):
-        return_obj.files.append(_ETreeXmlToObject.fill_instance_element(child, File))
-
-    for child in enum_results.findall('./Entries/Directory'):
-        return_obj.directories.append(
-            _ETreeXmlToObject.fill_instance_element(child, Directory))
-
-    for name, value in vars(return_obj).items():
-        if name == 'files' or name == 'directories':
-            continue
-        value = _ETreeXmlToObject.fill_data_member(enum_results, name, value)
-        if value is not None:
-            setattr(return_obj, name, value)
-
-    return return_obj
 
 
 def _create_file_result(response):